+++ tests/include/helperFunctions.h
@@ -8,7 +8,7 @@
 #pragma once
 
 #include "DeviceInfoMap.h"
-#include "RegisterInfoMap.h"
+#include "NumericAddressedRegisterCatalogue.h"
 
 void populateDummyDeviceInfo(ChimeraTK::DeviceInfoMap::DeviceInfo& deviceInfo, std::string dmapFileName,
     std::string deviceName = "card", std::string dev_file = "/dev/dummy_device_identifier",
